+++ src/common.ts
@@ -24,13 +24,19 @@ export const BPPAK_MODSFOLDER_PATH = path.join('Pal', 'Content', 'Paks', 'LogicM
 
 export const LUA_EXTENSIONS = ['.lua'];
 export const PAK_EXTENSIONS = ['.pak', '.utoc', '.ucas'];
-export const IGNORE_CONFLICTS = ['enabled.txt', 'ue4sslogicmod.info', '.ue4sslogicmod', '.logicmod'];
 
+export const MODS_FILE = 'mods.txt';
+export const MODS_FILE_BACKUP = 'mods.txt.original';
 export const UE4SS_ENABLED_FILE = 'enabled.txt';
+
+export const IGNORE_CONFLICTS = [UE4SS_ENABLED_FILE, 'ue4sslogicmod.info', '.ue4sslogicmod', '.logicmod'];
+export const IGNORE_DEPLOY = [MODS_FILE, MODS_FILE_BACKUP, UE4SS_ENABLED_FILE];
+
+export const UE4SS_DWMAPI = 'dwmapi.dll';
 export const XBOX_UE4SS_XINPUT_REPLACEMENT = 'xinput1_4.dll';
-export const UE4SS_FILES = [
-  'xinput1_3.dll', 'UE4SS-settings.ini',
-];
+export const UE4SS_SETTINGS_FILE = 'UE4SS-settings.ini';
+export const UE4SS_2_5_2_FILES = ['xinput1_3.dll', UE4SS_SETTINGS_FILE];
+export const UE4SS_3_0_0_FILES = [UE4SS_DWMAPI, UE4SS_SETTINGS_FILE];
 
 export const UE_PAK_TOOL_FILES = [
   'UnrealPak.exe',
@@ -42,13 +48,13 @@ export const TOP_LEVEL_DIRECTORIES = [
 
 export const MOD_TYPE_PAK = 'palworld-pak-modtype';
 export const MOD_TYPE_LUA = 'palworld-lua-modtype';
+export const MOD_TYPE_LUA_V2 = 'palworld-lua-modtype-v2';
 export const MOD_TYPE_BP_PAK = 'palworld-blueprint-modtype';
 export const MOD_TYPE_UNREAL_PAK_TOOL = 'palworld-unreal-pak-tool-modtype';
 
 export type PakModType = 'palworld-pak-modtype' | 'palworld-blueprint-modtype';
 
 export const UE4SS_XINPUT_FILENAME = 'UE4SS_v3.0.0.zip';
-export const UE4SS_DWMAPI = 'dwmapi.dll';
 export const UE_PAK_TOOL_FILENAME = 'UnrealPakTool.zip';
 
 export const PLUGIN_REQUIREMENTS: IPluginRequirement[] = [
@@ -58,7 +64,7 @@ export const PLUGIN_REQUIREMENTS: IPluginRequirement[] = [
     assemblyFileName: UE4SS_DWMAPI,
     userFacingName: 'UE4 Scripting System',
     githubUrl: 'https://api.github.com/repos/UE4SS-RE/RE-UE4SS',
-    findMod: (api: types.IExtensionApi) => findModByFile(api, '', UE4SS_FILES[1]),
+    findMod: (api: types.IExtensionApi) => findModByFile(api, '', UE4SS_SETTINGS_FILE),
     findDownloadId: (api: types.IExtensionApi) => findDownloadIdByPattern(api, PLUGIN_REQUIREMENTS[0]),
     fileArchivePattern: new RegExp(/^UE4SS.*v(\d+\.\d+\.\d+)/, 'i'),
     resolveVersion: (api: types.IExtensionApi) => resolveVersionByPattern(api, PLUGIN_REQUIREMENTS[0]),